feat: adiciona dtos para rating

+++ src/main/java/com/anthony/blacksmithOnlineStore/controler/dto/blacksmith/BlacksmithResponseDto.java
@@ -1,8 +1,9 @@
 package com.anthony.blacksmithOnlineStore.controler.dto.blacksmith;
 
 import com.anthony.blacksmithOnlineStore.entity.Blacksmith;
+import java.math.BigDecimal;
 
-public record BlacksmithResponseDto(Long id, String name, String description, Double rating) {
+public record BlacksmithResponseDto(Long id, String name, String description, BigDecimal rating) {
 
   public static BlacksmithResponseDto fromEntity(Blacksmith blacksmith) {
     return new BlacksmithResponseDto(
